fix #7818 feat(nimbus): Merge details page into summary
