+++ app/experimenter/nimbus-ui/src/components/AppLayoutSidebarLaunched/index.tsx
@@ -190,18 +190,6 @@ export const AppLayoutSidebarLaunched = ({
                 canReview={experiment.canReview}
               />
 
-              <li>
-                <ul className="text-dark list-unstyled ml-4">
-                  <LinkNav
-                    className="mx-1 mb-2 ml-3"
-                    storiesOf="pages/SummaryDetails"
-                    route={`${slug}/details`}
-                  >
-                    Details
-                  </LinkNav>
-                </ul>
-              </li>
-
               {analysisAvailable(analysis) ? (
                 <ResultsAvailableNav />
               ) : (
