+++ app/tests/integration/nimbus/test_experimenter.py
@@ -54,9 +54,8 @@ def test_promote_to_rollout(
     create_experiment,
 ):
     summary = create_experiment(selenium)
-    summary_detail = summary.navigate_to_details()
-    summary_detail.promote_first_branch_to_rollout()
-    summary_detail.wait_for_clone_parent_link_visible()
+    summary.promote_first_branch_to_rollout()
+    summary.wait_for_clone_parent_link_visible()
 
 
 @pytest.mark.run_once
